requirements.txt
pandas
numpy
requests
scikit-learn
lightgbm
catboost
matplotlib
plotly
folium
geopandas
osmnx
jupyter
pyarrow
pyyaml
tqdm
meteostat
pyyaml
pyarrow

